+++ CHANGES.md
@@ -2,3 +2,4 @@
 - 2019-12-04 v0.0.1 Added test.
 - 2019-12-04 v0.0.2 Triggering CI.
 - 2019-12-04 v0.0.3 Added badges and cleaned up godoc.
+- 2019-12-04 v0.0.4 Created a test case for the path-match Bad Pattern error.